Finish Task 1 - Added confirm to remove location

+++ ex/travel-tip-starter/js/app.controller.js
@@ -69,6 +69,9 @@ function renderLocs(locs) {
 }
 
 function onRemoveLoc(locId) {
+    const isConfirmed = confirm('Are you sure you want to remove this location?')
+    if (!isConfirmed) return
+
     locService.remove(locId)
         .then(() => {
             flashMsg('Location removed')
